Kernel interfaces initial check-in
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5047 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/kernels/streamset.cpp
@@ -33,7 +33,7 @@ llvm::Value * StreamSetBuffer::getBlockPointer(llvm::Value * blockNo) {
     if (mSegmentSize == 1) return mStreamSetBufferPtr;
     if (mSegmentSize == 0) return
         iBuilder->CreateGEP(getStreamSetBlockType(), mStreamSetBufferPtr, {blockNo});
-    Value * offset = iBuilder->CreateURem(blockNo, iBuilder->getInt32(mSegmentSize));
+    Value * offset = iBuilder->CreateURem(blockNo, iBuilder->getInt64(mSegmentSize));
     return iBuilder->CreateGEP(getStreamSetBlockType(), mStreamSetBufferPtr, {offset});
 }
